refactor(types): move ThemePalette to its own theme module; fix crate count
Judge debt items: ThemePalette sat in ids.rs (a file documented as
identifier types), and the architecture.md prose header still said 16
member crates while the tree in the same doc said 18.
