+++ parish/crates/parish-types/src/lib.rs
@@ -9,6 +9,7 @@ pub mod error;
 pub mod events;
 pub mod gossip;
 pub mod ids;
+pub mod theme;
 pub mod time;
 
 pub use conversation::{ConversationExchange, ConversationLog};
@@ -17,9 +18,10 @@ pub use error::ParishError;
 pub use events::{EventBus, GameEvent};
 pub use gossip::{GossipItem, GossipNetwork};
 pub use ids::{
-    DEFAULT_START_LOCATION, LanguageHint, Location, LocationId, NpcId, ThemePalette, Weather,
+    DEFAULT_START_LOCATION, LanguageHint, Location, LocationId, NpcId, Weather,
     extract_dialogue_from_partial_json, floor_char_boundary,
 };
+pub use theme::ThemePalette;
 pub use time::{
     DayType, Festival, GameClock, GameSpeed, Season, SpeedConfig, TimeOfDay, minute_word,
 };
